Feat: Set wallpaper (doesn't work)
The entry is written to the config file, but the setting doesn't take
effect and is overwritten with the default value in the config file
when logging onto the desktop for the first time. But the commented
out code is included because it shows how to write a value to a key
in a nested group.